+++ src/components/app.js
@@ -8,24 +8,8 @@ import AllThreads from './allthreads';
 import Navbar from './navbar';
 import NewPost from './newPost';
 import Sidebar from './sidebar';
-import Thread from './thread';
-import ThreadFeed from './threadFeed';
-
-// const App = () => {
-//     return (
-//             <div>
-//                 <Navbar/>
-//                 <div className="container">
-//                     <div className="row">
-//                         <Sidebar/>
-                     
-//                         <Route path='/newPost' component={NewPost}/>
-//                     </div>
-
-
-
-// import NewPost from './newPost';
-
+import Thread from './uniqueThread';
+// import ThreadFeed from './threadFeed';
 
 const UniqueThread = ({ match }) => {
     return(
@@ -40,12 +24,13 @@ const App = () => {
             <div className="container">
                 <div className="row">
                     <Sidebar/>
-                    <Route exact path="/" component={ThreadFeed} />
+                    {/* <Route exact path="/" component={ThreadFeed} /> */}
                     <Route exact path="/" component={AllThreads} />
                     <Route path='/newPost' component={NewPost}/>
-                    <Route path='/thread/:threadID' component={UniqueThread} />
+                    <Route path='/thread/:threadID' component={UniqueThread}  />
                 </div>
             </div>
+        </div>
     )
 }
